Moving trainsamplepage to simonsound; Bugfix in ssc / sscd: Mirophone / Soundcard config hung ssc Making soundcard / microphone ids 32 bit long Adding volumepage to sscs training if requested

+++ simonlib/sscobjects/microphone.cpp
@@ -21,7 +21,7 @@
 #include <QDataStream>
 #include <QByteArray>
 
-Microphone::Microphone(qint16 id, const QString& model, const QString& type) :
+Microphone::Microphone(qint32 id, const QString& model, const QString& type) :
 	m_id(id),
 	m_model(model),
 	m_type(type)
