+++ simonlib/sscobjects/soundcard.cpp
@@ -21,7 +21,7 @@
 #include <QDataStream>
 #include <QByteArray>
 
-SoundCard::SoundCard(qint16 id, const QString& model, const QString& type) :
+SoundCard::SoundCard(qint32 id, const QString& model, const QString& type) :
 	m_id(id),
 	m_model(model),
 	m_type(type)
