+++ simonlib/sscobjects/soundcard.h
@@ -27,23 +27,23 @@
 class SSCOBJECTS_EXPORT SoundCard : public SSCObject {
 
 private:
-	qint16 m_id;
+	qint32 m_id;
 	QString m_model;
 	QString m_type;
 
 public:
-	SoundCard(qint16 id, const QString& model, const QString& type);
+	SoundCard(qint32 id, const QString& model, const QString& type);
 
 	SoundCard() {}
 
 	void deserialize(QByteArray data);
 	QByteArray serialize();
 
-	qint16 id() { return m_id; }
+	qint32 id() { return m_id; }
 	QString model() { return m_model; }
 	QString type() { return m_type; }
 
-	void setId(qint16 id) { m_id = id; }
+	void setId(qint32 id) { m_id = id; }
 
 	~SoundCard() {}
 };
